@@ -4,23 +4,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-free-regular": "^5.0.13",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.19",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^26.0.24",
-    "@types/react": "^16.14.67",
-    "@types/react-dom": "^16.9.25",
-    "autosize": "^4.0.4",
+    "@fortawesome/fontawesome-free": "^7.1.0",
+    "@fortawesome/fontawesome-svg-core": "^7.1.0",
+    "@fortawesome/free-solid-svg-icons": "^7.1.0",
+    "@fortawesome/react-fontawesome": "^3.1.0",
     "firebase": "^12.4.0",
     "font-awesome": "^4.7.0",
-    "gh-pages": "^6.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
     "react-fontawesome": "^1.7.1",
-    "sweetalert2": "^10.16.11"
+    "sweetalert2": "^11.26.1",
+    "zustand": "^5.0.8"
   },
   "scripts": {
     "predeploy": "pnpm build",
@@ -29,6 +23,7 @@
     "build": "vite build",
     "preview": "vite preview",
     "prettier": "prettier . --write",
+    "typecheck": "tsc --noEmit",
     "test": "vitest"
   },
   "prettier": {},
@@ -49,12 +44,18 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.9.1",
-    "@types/autosize": "^3.0.7",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/autosize": "^4.0.3",
     "@types/node": "^24.7.2",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.1",
     "@vitejs/plugin-react": "^5.0.4",
     "@vitest/coverage-v8": "3.2.4",
+    "autosize": "^6.0.1",
+    "gh-pages": "^6.3.0",
     "prettier": "3.6.2",
-    "typescript": "^3.9.10",
+    "typescript": "^5.9.3",
     "vite": "^7.1.9",
     "vitest": "^3.2.4"
   }
