@@ -1,5 +1,5 @@
 import React from "react";
-import { describe, it } from "vitest";
+import { describe, expect, it } from "vitest";
 import { fireEvent, render, screen } from "../util/TestUtils";
 import SearchBar from "./SearchBar";
 
